refactor(compile): split native manifest resolver
